requirements.txt
alembic==0.9.10
asn1crypto==0.24.0
Brlapi==0.6.7
certifi==2018.1.18
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==6.7
cryptography==2.1.4
cupshelpers==1.0
debtags==2.1
Django==2.1.11
djangorestframework==3.10.2
dominate==2.3.1
dynamicbinder==0.1
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.2.1
Flask-Moment==0.6.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
httplib2==0.9.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
keyring==10.6.0
keyrings.alt==3.0
louis==3.5.0
Mako==1.0.7
MarkupSafe==1.0
ply==3.11
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
pygobject==3.28.2
pysmbc==1.0.15.6
python-apt==1.6.0
python-dateutil==2.7.3
python-debian==0.1.32
python-debianbts==2.7.2
python-editor==1.0.3
pytz==2018.9
pyxattr==0.6.0
pyxdg==0.25
reportbug==7.1.10
requests==2.18.4
SecretStorage==2.3.1
six==1.11.0
SQLAlchemy==1.2.9
unattended-upgrades==0.1
urllib3==1.22
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
youtube-dl==2018.4.25
